+++ _printf.c
@@ -1,62 +1,28 @@
 #include "main.h"
-
-void print_buffer(char buffer[], int *buff_ind);
-
-/**
- * _printf - func that produces output according to a format
- * @format: the string that holds any format specifier
- * Return: the number of printed characters
- */
 int _printf(const char *format, ...)
 {
-	int i, printed = 0, printed_chars = 0;
-	int flags, width, precision, size, buff_ind = 0;
-	va_list list;
-	char buffer[BUFFER_SIZE];
+	int i = 0, len = 0;
 
-	if (format == NULL)
-		return (-1);
+	char *buffer;
+
+	va_list list;
 
 	va_start(list, format);
 
-	for (i = 0; format && format[i] != '\0'; i++)
-
+	while (format[i] != '\0')
 	{
-		if (format[i] != '%')
+		if (format[i] ==  '%')
 		{
-			buffer[buff_ind++] = format[i];
-			if (buff_ind == BUFFER_SIZE)
-				print_buffer(buffer, &buff_ind);
-			printed_chars++;
+			len = handle_print(format, list, buffer, i);
+			if (len == 1)
+			{
+				i += 2;
+				continue;
+			}
 		}
-		else
-		{
-			print_buffer(buffer, &buff_ind);
-			flags = get_flags(format, &i);
-			width = get_width(format, &i, list);
-			precision = get_precision(format, &i, list);
-			size = get_size(format, &i);
-			i++;
-			printed = handle_print(format, &i, list, buffer, flags, width, precision, size);
-			if (printed == -1)
-				return (-1);
-			printed_chars += printed;
-		}
+			     _putchar(format[i]);
+		i++;
 	}
-		print_buffer(buffer, &buff_ind);
-
-		va_end(list);
-
-		return (printed_chars);
-}
-
-/** print_buffer - Prints the contents of the buffer
- * @buffer: array of chars
- * @buff_ind: Index at which to add next char
- */
-void print_buffer(char buffer[], int *buff_ind)
-{
-	if (*buff_ind > 0)
-		write(1, &buffer[0], *buff_ind);
-	*buff_ind = 0;
+	va_end(list);
+	return (1);
 }
