+++ backup/get_flags.c
