+++ backup/get_width.c
