+++ backup/handle_print.c
@@ -13,7 +13,7 @@ int handle_print(const char *fmt, va_list list,  char *buffer, int curr_i)
 	int i;
 
 	typ_t type_ops[] = {
-		{'c', prints},
+		{'c', print},
 		{'s', prints},
 	};
 	for (i = 0; i < 2; i++)
@@ -23,6 +23,11 @@ int handle_print(const char *fmt, va_list list,  char *buffer, int curr_i)
 
 		else if (fmt[curr_i + 1] == '\0')
 			return (-1);
+		else if (fmt[curr_i] & fmt[curr_i + 1] == '%')
+		{
+			_putchar('%');
+			return (1);
+		}
 		else
 		{
 			_putchar(fmt[curr_i]);
