+++ backup/new_get_func.c
