+++ backup/print_nstring.c~
