+++ new/handle_print.c
@@ -15,18 +15,15 @@ int handle_print(const char *fmt, va_list list,  char *buffer, int curr_i)
 	typ_t type_ops[] = {
 		{'c', prints},
 		{'s', prints},
+		{'d', _printdec},
 	};
-	for (i = 0; i < 2; i++)
+	for (i = 0; i < 3; i++)
 	{
 		if (fmt[curr_i] == '%' && fmt[curr_i + 1] == type_ops[i].fmt)
 			return (type_ops[i].func(fmt, list, buffer, curr_i));
 
 		else if (fmt[curr_i + 1] == '\0')
 			return (-1);
-		else
-		{
-			continue;
-		}
 	}
 	return (1);
 }
